+++ lib/src/binance/models/models.dart
@@ -1,3 +1,4 @@
+export 'aggregated_trade/aggregated_trade.dart';
 export 'exchange_filter/exchange_filter.dart';
 export 'exchange_info/exchange_info.dart';
 export 'order_book/order_book.dart';